refactor: userRepository store List-> Map으로 변경
- UserRepository store를 List -> TreeMap으로 변경
- User Profile을 보여주기 위한 정보를 ProfileViewDTO를 통해 표현
- User를 식별하기 위한 기본키인 id를 추가

+++ src/main/java/com/kakao/cafe/user/domain/User.java
@@ -8,7 +8,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class User {
 
-    private int id;
+    private Long id;
 
     private String userId;
 
@@ -18,7 +18,7 @@ public class User {
 
     private String email;
 
-    public User(int id, SignUpDTO signUpDTO) {
+    public User(Long id, SignUpDTO signUpDTO) {
         this.id = id;
         this.userId = signUpDTO.getUserId();
         this.password = signUpDTO.getPassword();
